Time: 19 ms (47.99%) | Memory: 19.7 MB (9.86%) - LeetSync

+++ 283-move-zeroes/move-zeroes.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     void moveZeroes(vector<int>& nums) {
      short int zc=0,p=0,s=nums.size();
-        for(int i=0;i<s;i++)
+        for(short int i=0;i<s;i++)
         {
             if(nums[i]==0)
                 ++zc;
